# refactor(assembly): speed up script compilation