--- a/core/src/program/blocks/span_block.rs
+++ b/core/src/program/blocks/span_block.rs
@@ -91,17 +91,6 @@ impl Span {
         }
         Self::new(ops)
     }
-
-    /// Appends the operations from the provided [Span] to this [Span].
-    pub fn append(&mut self, other: Self) {
-        let mut ops = self.get_ops();
-        for batch in other.op_batches {
-            ops.extend_from_slice(&batch.ops);
-        }
-        let (op_batches, hash) = batch_ops(ops);
-        self.op_batches = op_batches;
-        self.hash = hash;
-    }
 }
 
 impl fmt::Display for Span {