This PR address https://github.com/maticnetwork/miden/issues/153.

This PR significantly reduces the script compilation time by a factor of 36. 

The following benchmark were conducted on AMD Ryzen 5 3550H running Pop!_OS 22.04 LTS:
Before this commit: 
`sha256:  [2.8630 s 2.8750 s 2.8874 s]`

Post this commit:
`sha256:  [78.123 ms 78.921 ms 79.821 ms]`

The reason for this big increase in the speed up is because now instead of iteratively combining two consecutive SPAN blocks(and applying `hash elements`), we are keeping track of all the consecutive SPAN blocks occured in a row and apply `hash_elements` only once when there's a discontinuity. That way we have to apply the `hash_elements` only once v/s every single time we get a pair of span blocks in the earlier case.