@bobbinth 2022-06-17
One other thing I'm curious about: does this noticeably reduce runtime of tests. Meaning, if you run `cargo test` before and after this change - do you see any improvement?