@0xKanekiKen 2022-06-17
@bobbinth: Thanks for review! 

`Not for this PR (but maybe we should create an issue for this): There is probably a more efficient way to handle this. `
I've created a new issue for this. You can assign it to me. #253 

`One other thing I'm curious about: does this noticeably reduce runtime of tests. Meaning, if you run cargo test before and after this change - do you see any improvement?`
In terms of speed boost, we are not getting much. Earlier the integration tests(rest all are pretty quick) were taking ~440 secs and now post this commit they are taking ~424 secs on my system. There are few tests(`div_fail`, `lt`, `cdrop`, `dupw`, `u32_rotr_b` etc.) which are still taking considerable amount of time to run.
We should first address #253 and then see if it helps with other tests as I see it the ideal next step post this commit. Love to hear your thoughts on this.