@bobbinth 2022-06-17
> In terms of speed boost, we are not getting much. Earlier the integration tests(rest all are pretty quick) were taking ~440 secs and now post this commit they are taking ~424 secs on my system.

Makes sense. One more question: we have a few tests which are being ignored now (test for SHA256 being one of them). Could you check what is the runtime of these tests before and after this PR? If the runtime is small now, we might as well turn them on.