@0xKanekiKen 2022-06-18
`Makes sense. One more question: we have a few tests which are being ignored now (test for SHA256 being one of them). Could you check what is the runtime of these tests before and after this PR? If the runtime is small now, we might as well turn them on.`

2 tests were being ignored at the moment: `keccak256_bit_interleaving` & `keccak256_2_to_1_hash`. When I included them in the integration tests, the runtime I got on my machine is as follows:

Post commit: `430.62s`
Pre commit: `602.36s`

Based on the above finding I think it's fine if we turn them on. I've turned them on in this PR.